refactor: useTodos 훅 모듈화
- 필터 상수를 filters.js 파일로 분리
- 핸들러 함수를 todoHandlers.js 파일로 분리
