+++ src/hooks/useTodos.js
@@ -1,33 +1,15 @@
 import { useState } from "react";
-import addTodo from "../todoUtils/addTodo";
-import toggleComplete from "../todoUtils/toggleComplete";
-import deleteTodo from "../todoUtils/deleteTodo";
-
-const FILTERS = {
-  ALL: "all",
-  ACTIVE: "active",
-  COMPLETED: "completed",
-};
+import FILTERS from "../todoUtils/filters";
+import {
+  handleAddTodo,
+  handleToggleComplete,
+  handleDeleteTodo,
+  handleSetFilter,
+} from "../todoUtils/todoHandlers";
 
 const useTodos = () => {
   const [todos, setTodos] = useState([]);
   const [filter, setFilter] = useState(FILTERS.ALL);
-
-  const handleAddTodo = (todo) => {
-    setTodos((prevTodos) => addTodo(prevTodos, todo));
-  };
-
-  const handleToggleComplete = (index) => {
-    setTodos((prevTodos) => toggleComplete(prevTodos, index));
-  };
-
-  const handleDeleteTodo = (index) => {
-    setTodos((prevTodos) => deleteTodo(prevTodos, index));
-  };
-
-  const handleSetFilter = (filter) => {
-    setFilter(filter);
-  };
 
   const filteredTodos = todos.filter((todo) => {
     if (filter === FILTERS.ACTIVE) return !todo.completed;
@@ -37,10 +19,11 @@ const useTodos = () => {
 
   return {
     todos: filteredTodos,
-    handleAddTodo,
-    handleToggleComplete,
-    handleDeleteTodo,
-    handleSetFilter,
+    handleAddTodo: (todo) => handleAddTodo(todos, setTodos, todo),
+    handleToggleComplete: (index) =>
+      handleToggleComplete(todos, setTodos, index),
+    handleDeleteTodo: (index) => handleDeleteTodo(todos, setTodos, index),
+    handleSetFilter: (filter) => handleSetFilter(setFilter, filter),
     filter,
     FILTERS,
   };
