+++ src/todoUtils/filters.js
@@ -1,0 +1,7 @@
+const FILTERS = {
+  ALL: "all",
+  ACTIVE: "active",
+  COMPLETED: "completed",
+};
+
+export default FILTERS;
